Move ffmpeg, node resolution funcs to utils
